moved CardContainer.js
